backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
alembic==1.13.1
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.26.0
structlog==24.1.0
aiogram>=3.27.0
aiosmtplib>=5.1.0
psycopg2-binary==2.9.9
websockets==12.0
